refactor(ui): :recycle: move dialogs from main.js to individual rigs
